# pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: Benchmarkoor API
  description: |
    API for the Benchmarkoor Ethereum execution client benchmarking tool.

    ## Authentication

    Most endpoints require authentication via either:
    - **Session cookie** (`benchmarkoor_session`) — obtained by logging in via `/auth/login` or GitHub OAuth
    - **Bearer token** — an API key prefixed with `bmk_`, passed in the `Authorization: Bearer <key>` header

    Endpoints under `/admin` additionally require the `admin` role.

    ## PostgREST-style Query Endpoints

    The `/index/query/*` endpoints support PostgREST-style filtering:
    - **Filtering:** Add query parameters in the form `column=operator.value` (e.g. `client=eq.geth`, `tests_total=gt.10`)
    - **Operators:** `eq`, `neq`, `gt`, `gte`, `lt`, `lte`, `like`, `in`, `is`
    - **Sorting:** `order=column.asc` or `order=column.desc` (comma-separated for multiple)
    - **Pagination:** `limit` and `offset`
    - **Column selection:** `select=col1,col2`
  version: 1.0.0
servers:
  - url: /api/v1
...
security: []
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a user
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    put:
      operationId: updateUser
      tags: [admin]
      summary: Update a user
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/sessions:
    get:
      operationId: listSessions
      tags: [admin]
      summary: List all active sessions
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: All active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/api-keys/{id}:
    delete:
      operationId: deleteAPIKey
      tags: [admin]
      summary: Delete any API key
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: API key deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: benchmarkoor_session
    bearerAuth:
      type: http
      scheme: bearer
      description: API key prefixed with `bmk_`
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    StatusOk:
      type: object
      properties:
        status:
          type: string
          example: ok

    # ── Auth schemas ────────────────────────────────────────────────
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, readonly]
        source:
          type: string
          enum: [config, admin, github]
...
    SessionResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        username:
          type: string
        source:
          type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        last_active_at:
          type: string
          format: date-time
          nullable: true

    # ── Admin schemas ───────────────────────────────────────────────
    CreateUserRequest:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, readonly]
    UpdateUserRequest:
      type: object
      properties:
        password:
          type: string
        role:
          type: string
          enum: [admin, readonly]
...